+++ .changeset/soft-tempo-chains.md
@@ -1,0 +1,5 @@
+---
+'mppx': patch
+---
+
+Updated Tempo chain imports to use the `viem/tempo/chains` entrypoint.
