+++ examples/charge/package.json
@@ -15,7 +15,7 @@
     "bun": "^1.3.13",
     "mppx": "latest",
     "typescript": "~6.0.3",
-    "viem": "^2.47.6",
+    "viem": "^2.50.4",
     "vite": "^8.0.10"
   }
 }
