+++ package.json
@@ -51,7 +51,7 @@
     "testcontainers": "^11.14.0",
     "tsx": "^4.21.0",
     "typescript": "~6.0.3",
-    "viem": "^2.47.6",
+    "viem": "^2.50.4",
     "vite": "^8.0.10",
     "vp": "npm:vite-plus@~0.1.17",
     "ws": "^8.20.0",
