+++ pnpm-workspace.yaml
@@ -14,7 +14,7 @@ overrides:
   vitest: 'npm:@voidzero-dev/vite-plus-test@~0.1.17'
   typescript: '~5.9.3'
   ox: '0.14.18'
-  viem: '^2.47.5'
+  viem: '^2.50.4'
   path-to-regexp@<8.4.0: '8.4.0'
   tar@<=7.5.10: '7.5.11'
   '@modelcontextprotocol/sdk@>=1.10.0 <=1.25.3': '1.26.0'
