+++ src/tempo/client/Session.ts
@@ -1,6 +1,6 @@
 import type { Hex } from 'ox'
 import { type Address, parseUnits, type Account as viem_Account } from 'viem'
-import { tempo as tempo_chain } from 'viem/chains'
+import { tempo as tempo_chain } from 'viem/tempo/chains'
 
 import type * as Challenge from '../../Challenge.js'
 import * as Method from '../../Method.js'
